requirements.txt
cassandra-driver==3.24.0
click==7.1.2
geomet==0.2.1.post1
numpy==1.19.2
pandas==1.1.2
pyarrow==1.0.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
